Use int everywhere

+++ gtk2_ardour/tracker/track_pattern.cc
@@ -151,31 +151,31 @@ TrackPattern::audio_track_pattern ()
 }
 
 Temporal::Beats
-TrackPattern::region_relative_beats (uint32_t rowi, int mri, int32_t delay) const
+TrackPattern::region_relative_beats (int rowi, int mri, int delay) const
 {
 	return Temporal::Beats ();
 }
 
 int64_t
-TrackPattern::region_relative_delay_ticks (const Temporal::Beats& event_time, uint32_t rowi, int mri) const
+TrackPattern::region_relative_delay_ticks (const Temporal::Beats& event_time, int rowi, int mri) const
 {
 	return 0;
 }
 
 bool
-TrackPattern::is_auto_displayable (uint32_t rowi, int mri, const Evoral::Parameter& param) const
+TrackPattern::is_auto_displayable (int rowi, int mri, const Evoral::Parameter& param) const
 {
 	return is_displayable (rowi, param);
 }
 
 size_t
-TrackPattern::automation_list_count (uint32_t rowi, int mri, const Evoral::Parameter& param) const
+TrackPattern::automation_list_count (int rowi, int mri, const Evoral::Parameter& param) const
 {
 	return param_to_row_to_ali.find (param)->second.count (rowi);
 }
 
 Evoral::ControlEvent*
-TrackPattern::get_automation_control_event (uint32_t rowi, int mri, const Evoral::Parameter& param) const
+TrackPattern::get_automation_control_event (int rowi, int mri, const Evoral::Parameter& param) const
 {
 	ParamToRowToAutomationListIt::const_iterator pit = param_to_row_to_ali.find (param);
 	if (pit == param_to_row_to_ali.end())
@@ -195,19 +195,19 @@ TrackPattern::is_region_defined (int rowi) const
 }
 
 int
-TrackPattern::to_rrri (uint32_t rowi, int mri) const
+TrackPattern::to_rrri (int rowi, int mri) const
 {
-	return (int)rowi;
+	return rowi;
 }
 
 int
-TrackPattern::to_rrri (uint32_t rowi) const
+TrackPattern::to_rrri (int rowi) const
 {
-	return (int)rowi;
+	return rowi;
 }
 
 int
-TrackPattern::to_mri (uint32_t rowi) const
+TrackPattern::to_mri (int rowi) const
 {
 	return -1;
 }
